admin-service/requirements.txt
asgiref==3.7.2
cffi==1.16.0
cryptography==41.0.5
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
mysqlclient==2.2.0
pycparser==2.21
PyJWT==2.8.0
python-dotenv==1.0.0
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
